+++ src/Event/Fcl/fclShopMisc.c
@@ -265,7 +265,7 @@ u32 FUN_003f2f70_i(int param_1,int param_2,int param_3);
  u32 FUN_003c58f0_i(u32 param_1,int param_2,u32 param_3,u32 param_4);
  #pragma alias FUN_0019d3f0_fcl FUN_0019d3f0
  void FUN_0019d3f0_fcl(const char *param_1,int param_2);
-u64 FUN_003f33d0(int param_1,u32 param_2);
+u32 FUN_003f33d0(int param_1,u32 param_2);
 u64 FUN_003f3970(long param_1,u32 param_2);
 u64 FUN_003f39a0(long param_1,long param_2,u32 param_3);
 void FUN_003f3dc0(int param_1,u64 param_2);
@@ -3559,7 +3559,7 @@ ret:
 // FUN_003F33D0 NONMATCHING
 
 
-u64 FUN_003f33d0(int param_1,u32 param_2)
+u32 FUN_003f33d0(int param_1,u32 param_2)
 {
   u32 option_flags;
   u32 task;
