+++ src/nmSequence.c
@@ -583,7 +583,7 @@ void FUN_003c58e0(void)
 // FUN_003C58F0 NONMATCHING
 
 
-u64 FUN_003c58f0(u32 param_1,int param_2,u32 param_3,u32 param_4)
+u64 FUN_003c58f0(u32 param_1,long param_2,u32 param_3,u32 param_4)
